restruct storage mod
