+++ rust/xaynet-server/src/storage/coordinator_storage/mod.rs
@@ -1,0 +1,1 @@
+pub mod redis;
