+++ rust/xaynet-server/src/storage/mod.rs
@@ -1,14 +1,13 @@
-pub mod api;
-pub mod noop_model_store;
-pub mod redis;
-#[cfg(feature = "model-persistence")]
-pub mod s3;
+pub mod coordinator_storage;
+pub mod model_storage;
 pub mod store;
 #[cfg(test)]
 pub(crate) mod tests;
+pub mod traits;
 
 pub use self::{
-    api::{
+    store::Store,
+    traits::{
         CoordinatorStorage,
         LocalSeedDictAdd,
         LocalSeedDictAddError,
@@ -20,7 +19,4 @@ pub use self::{
         SumPartAdd,
         SumPartAddError,
     },
-    noop_model_store::NoOpModelStore,
-    redis::{RedisError, RedisResult},
-    store::Store,
 };
