+++ rust/xaynet-server/src/storage/model_storage/mod.rs
@@ -1,0 +1,3 @@
+pub mod noop;
+#[cfg(feature = "model-persistence")]
+pub mod s3;
