+++ rust/xaynet-server/src/storage/traits.rs
